Brick Metabox controls now load using ajax.
In MP Core 1.0.2.1, the metabox creation class allows for metabox contents to be loaded-in via ajax. The changes in this commit cause the controls for this metabox to use that functionality and will now load using ajax when chosen by the user as a Content-Type in MP Stacks.

Scripts now enqueued only when needed.